Allow opening keymaps starting with `core/`
This will be used in https://github.com/atom/atom/pull/14024 where
keymaps are loaded during snapshot. Since we are unable to supply an
absolute path when that code runs, we will instead use `core/` in order
to allow this package to understand if it needs to extract a bundled
keymap from Atom's package.json.